Refactor docs into folders to support Docs plugin usage
This is a proposed new format for the docs that follows a more logical structure and more closely matches our URL structure for the docs.

It will enable people editing the docs locally (or via GitHub) to be able to follow links, which was previously not possible.

It should hopefully keep the history of changes as well.